bin/go.sh, bin/zig.sh: Use global and session modes for environment activation

+++ bin/go.sh
@@ -19,9 +19,10 @@ _GO_SRC="${GO_SRC:-"$_GOSH_CACHE/src"}"
 _main() {
 	local _action="activate"
 	local _mode="src"
+	local _global=""
 
-	local _opts="bcdhiLlps"
-	local _lopts="bin,clear,deinit,delete,help,init,list,prefix,releases,source"
+	local _opts="bcdghiLlps"
+	local _lopts="bin,clear,deinit,delete,global,help,init,list,prefix,releases,source"
 
 	eval set -- "$(
 		getopt --options "$_opts" --longoptions "$_lopts" --name "$0" -- "$@"
@@ -40,6 +41,11 @@ _main() {
 
 		-d | --delete)
 			_action="delete"
+			;;
+
+		-g | --global)
+			_action="activate"
+			_global="true"
 			;;
 
 		-h | --help)
@@ -185,8 +191,16 @@ _activate() {
 	fi
 
 	log.sh "activating release $_rel.."
-	rm -f "$_GOSH_DATA/go"
-	ln -s "$_env" "$_GOSH_DATA/go"
+
+	if [ -z "$_global" ]; then
+		echo "export PATH=$_env:$PATH"
+	else
+		log.sh "using global mode.."
+		rm -f "$_GOSH_DATA/go"
+		ln -s "$_env" "$_GOSH_DATA/go"
+		echo "export GOPATH=$_GOPATH"
+		echo "export PATH=$_GOPATH/bin:$_GOSH_DATA/go/bin:$PATH"
+	fi
 }
 
 _build() {
@@ -314,7 +328,13 @@ _host_os() {
 }
 
 _latest_release() {
-	local _releases="$(_releases)"
+	local _releases="$(_releases || echo "")"
+
+	if [ -z "$_releases" ]; then
+		log.sh "using local latest release.."
+		_releases="$(ls "$_GOSH_ENVS" | grep -v "release" | tail -n 1)"
+	fi
+
 	echo "$_releases" | head -n 1
 }
 
@@ -329,7 +349,7 @@ _show_help() {
 	cat << EOF
 $_name - manage Go environments.
 
-Usage: $_name [-b | -s] [RELEASE]
+Usage: $_name [-g] [-b | -s] [RELEASE]
    or: $_name -L
    or: $_name -l
    or: $_name -p [RELEASE]
@@ -343,6 +363,7 @@ Options:
   -b, --bin        Download pre-built binaries for the given release
 	-c, --clear      Clear cache data (build cache, binary downloads, etc...)
   -d, --delete     Remove the given release
+  -g, --global     Use given release as global version
   -h, --help       Show this help message
   -i, --init       Setup the environment for using $_name
   -L, --releases   List available releases
